requirements.txt
asgiref==3.8.1
black==24.8.0
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
Django==5.1.1
djangorestframework==3.15.2
filelock==3.15.4
gunicorn==23.0.0
identify==2.6.0
iniconfig==2.0.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.8.0
pytest==8.3.2
pytest-django==4.9.0
PyYAML==6.0.2
sqlparse==0.5.1
virtualenv==20.26.3
